+++ cx-portal/src/components/overlays/AppMarketplaceRequest/index.tsx
@@ -164,6 +164,7 @@ export default function AppMarketplaceRequest({ id }: { id: string }) {
           color={'primary'}
           label={t('content.appdetail.subscribe')}
           buttonData={OrderStatusButtonData}
+          selectable={false}
         />
       </DialogContent>