+++ CHANGELOG.md
@@ -24,6 +24,9 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
    * Do not hide current on zoom in action
 * Admin board - Service
    * Filter action issue fix
+* Appdetails
+   * App Images size and responsivenes
+   * Subscription Status hover changes
 
 ## 1.4.0